audio waveform todo

+++ app/src/main/java/com/paranid5/crescendo/domain/services/video_cash_service/CashTrimRange.kt
@@ -4,4 +4,4 @@ import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
-data class CashTrimRange(val offset: Int, val endPoint: Int) : Parcelable
+data class CashTrimRange(val offset: Long, val endPoint: Long) : Parcelable
